app/requirements.txt
Django==6.0
beautifulsoup4==4.14.3
django-environ==0.12.0
django-filter==25.2
django-tinymce==5.0.0
django-user-g11n==0.4.4
djangorestframework==3.16.1
djangorestframework-csv==3.0.2
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.12.1
fontawesomefree==6.6.0
pillow==12.0.0
psycopg[binary]==3.3.1
python-dateutil==2.9.0.post0
